Moved table to typed
